+++ src/components/dropdowns/dropdown.js
@@ -33,7 +33,7 @@ var _dropdownButton = require("../../components/dropdowns/dropdown-button");
 
 var _elementHelper = _interopRequireDefault(require("../../helpers/dom/element-helper"));
 
-function _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+function _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
 
 function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
 
@@ -168,7 +168,7 @@ var Dropdown = /*#__PURE__*/function (_React$Component) {
         top: 0
       };
       var arrowHorizontalPosition = this.props.arrowHorizontalPosition;
-      return (0, _extends2.default)((0, _extends2.default)({}, defaultPosition), arrowHorizontalPosition);
+      return (0, _extends2.default)({}, defaultPosition, arrowHorizontalPosition);
     }
   }, {
     key: "getContentPosition",
@@ -177,7 +177,7 @@ var Dropdown = /*#__PURE__*/function (_React$Component) {
         top: this.state.referenceElementHeight
       };
       var horizontalPosition = this.props.contentHorizontalPosition;
-      return (0, _extends2.default)((0, _extends2.default)({}, defaultPosition), horizontalPosition);
+      return (0, _extends2.default)({}, defaultPosition, horizontalPosition);
     } // Component listener callbacks
 
   }, {